refactor rpc common code and move into a common util
