+++ trinity/rpc/json_rpc/modules/net.py
@@ -3,7 +3,7 @@ from lahja import EndpointAPI
 from trinity.constants import TO_NETWORKING_BROADCAST_CONFIG
 from trinity.nodes.events import NetworkIdRequest
 from trinity.protocol.common.events import PeerCountRequest
-from trinity.rpc.modules import BaseRPCModule
+from trinity.rpc.json_rpc.modules import BaseRPCModule
 
 
 class Net(BaseRPCModule):
